requirements.txt
# LLM Providers
anthropic>=0.40.0
google-genai>=1.0.0

# Gmail API
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0

# Telegram
requests>=2.31

# Web framework (webhook mode)
fastapi>=0.110.0
uvicorn>=0.27.0

# Scheduler
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0
tenacity>=8.2.0
segno>=1.6.6
phonenumbers>=8.13.0
tavily-python>=0.5.0
duckduckgo-search>=6.3.0
httpx>=0.28.0
cryptography>=43.0.0

# Work Email (IMAP)
pdfplumber>=0.10.0
python-docx>=1.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
